Switch over to using getrusage() for the CPU time.
For some reason, clock() returns a value on OS X that's off by a factor
ten thousand. I don't feel like debugging this in depth, so let's just
go ahead and switch over to getrusage(), which provides the same
functionality. The nice thing about this function is that we know it has
microsecond precision, as it uses a struct timeval.