Make poll()/select() work (better) within the emulator.
After processing the futex/clock-related patterns of subscriptions
passed to poll(), fall back to polling on file descriptors. We can do
this by referencing all of the file descriptors listed in the
subscriptions and turning those into struct pollfds. After the poll
call, we turn the pollfds back into events.