Reimplement cloudabi-run in C++.
Now that we're using C++, we can make use of the yaml2argdata code, so
that YAML parsing as done by cloudabi-run and Scuba will remain in sync
(with the exception of file descriptor specific tags).

For now, only support !fd and !file. We're not going to re-add !socket,
as we'll add a special tag to construct Flower switchboard channels
instead.